chore: Rename livekit > meetings provider
